Separate main page from index
